Replace #define syscalls cancel machinery.
Use inline functions for type-checking. Separate CL will allow
user to redefine/catch these functions.

R=pasko@google.com

Review URL: http://codereview.chromium.org/7785030

+++ sysdeps/nacl/read.c
@@ -8,7 +8,7 @@
 
 ssize_t __libc_read (int fd, void *buf, size_t size)
 {
-  int nread;
+  size_t nread;
   int result = __nacl_irt_read (fd, buf, size, &nread);
   if (result != 0) {
     errno = -result;
